add change role

+++ _actions/delete.php
@@ -7,7 +7,12 @@ use Helpers\Auth;
 use Libs\Database\MySQL;
 use Libs\Database\UsersTable;
 
-Auth::check();
+$auth = Auth::check();
+    
+if(!$auth|| $auth->role_id != 2) {
+    HTTP::redirect("admin/login.php", "auth=fail");
+    exit();
+}
 
 $table = new UsersTable(new MySQL);
 $table->delete($_GET["id"]);
